Make resulting output file read-only
add an optional argument in wrap_close to make
the file read-only. This also made making the list of subroutines
in the convert_mksrf program under "contains" so that an explicit
argument list could be used, which then required that scalar wrap_def_var
calls be sent a dimension variable array (rather than just 0).

Also update the namelist with the latest files, and also make sure
the output file names aren't outputing over the input files.